requirements.txt
psycopg2-binary
Django
gunicorn
whitenoise
django-cors-headers
requests
cloudinary
django-phonenumber-field
django-cloudinary-storage
djangorestframework-simplejwt
djangorestframework
supabase